# [dv] Improve sim performace for tl_errors