@sriyerg 2020-01-28
Thanks, 530s to 7s is awesome speedup!