@weicaiyang 2020-01-30
@udinator 
As discussed, good coding style on constraint can save significant simulation time and we can add following in our DV style guide for constraint.
1. Avoid using loop in constraint if possible. In some cases, `foreach` can replaced by `inside`

:-1:
```systemverilog
constraint con{
  foreach (data[i]) {
    a != data[i];
  }
}
```
:+1: 
```systemverilog
constraint con{!(a inside {data});}
```

2. Avoid doing calculation in loop in constraint if possible
3. Where possible, prefer mask operator (a & mask) over modulus (a % size) operator, as mask operator is much faster.
4. When constraint is very complex, recommend to split those relevant variables and randomize them separately.