@udinator 2020-01-30
> @udinator
> As discussed, good coding style on constraint can save significant simulation time and we can add following in our DV style guide for constraint.
> 
> 1. Avoid using loop in constraint if possible. In some cases, `foreach` can replaced by `inside`
> 
> 
> 
> ```systemverilog
> constraint con{
>   foreach (data[i]) {
>     a != data[i];
>   }
> }
> ```
> 
> 
> 
> ```systemverilog
> constraint con{!(a inside {data});}
> ```
> 
> 1. Avoid doing calculation in loop in constraint if possible
> 2. Where possible, prefer mask operator (a & mask) over modulus (a % size) operator, as mask operator is much faster.
> 3. When constraint is very complex, recommend to split those relevant variables and randomize them separately.

@weicai, thanks for compiling a brief list! I agree with everything you say here, and I'll add a section to the style guide about this.
One addition that I would suggest is that when randomizing an array of objects, do not directly randomize the array, but instead iterate through it and randomize each element of the array, either in `post_randomize()` if the array is a rand member of a cfg object or some parent object that is getting randomized, or directly access the array if this is not the case.